mongodb queries: Update api and response queries to gat data from mongodb.
