+++ src/data/serverData.js
@@ -1,6 +1,6 @@
-import {mongoDb, ObjectId} from './mongoData.js';
 import fetch from 'node-fetch';
 import dotenv from './environment.js';
+
 class ServerData {
   
   constructor () {
@@ -68,54 +68,6 @@ class ServerData {
       console.log(err);
       return { status: 500, message : "error" };
     });
-  }
-
-  async getApiData () {
-    let apiData = mongoDb.getApis();
-
-    if (apiData) {
-      let apis = await apiData.find({}).toArray();
-      return apis;
-    }
-    else {
-      return { status: 404, message : "error" };
-    }
-  }
-
-  async getApiById (apiId) {
-    let apiData = mongoDb.getApis();
-
-    if (apiData) {
-      let api = await apiData.findOne(ObjectId(apiId));
-      return api;
-    }
-    else {
-      return { status: 404, message : "error" };
-    }
-  }
-
-  async getResponseData () {
-    let responseData = mongoDb.getResponses();
-
-    if (responseData) {
-      let responses = await responseData.find({}).toArray();
-      return responses;
-    }
-    else {
-      return { status: 404, message : "error" };
-    }
-  }
-
-  async getResponseById (responseId) {
-    let responseData = mongoDb.getResponses();
-
-    if (responseData) {
-      let res = await responseData.findOne(ObjectId(responseId));
-      return res;
-    }
-    else {
-      return { status: 404, message : "error" };
-    }
   }
 }
