+++ src/resolvers/responses/queries.js
@@ -1,12 +1,28 @@
-import { ServerData } from '../../data/serverData.js';
-const serverData = new ServerData();
+import {mongoDb, ObjectId} from '../../data/mongoData.js';
 
 const responseQueries = {
-  response: (parent, args) => {
-    return serverData.getResponseById(args._id);
+  response: async (parent, args) => {
+    let responseId = args._id;
+    let responseData = mongoDb.getResponses();
+
+    if (responseData) {
+      let res = await responseData.findOne(ObjectId(responseId));
+      return res;
+    }
+    else {
+      return { status: 404, message : "error" };
+    }
   },
-  responses: () => {
-    return serverData.getResponseData();
+  responses: async () => {
+    let responseData = mongoDb.getResponses();
+
+    if (responseData) {
+      let responses = await responseData.find({}).toArray();
+      return responses;
+    }
+    else {
+      return { status: 404, message : "error" };
+    }
   }
 };
